requirements.txt
Flask==3.1.2
Pillow==11.3.0
numpy==2.3.3
Werkzeug==3.1.3
